# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Staging Superapp API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server (Docker)"
        }
    ],
    "paths": {
...
        "/api/v1/purchase/orders": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Get list of purchase orders",
                "operationId": "89cbb8f0d2a7701962038ee59e78ae84",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[supplier_id]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Create a new Purchase Order",
                "operationId": "d01bb6ad4d00462721bb26c0e8aefd9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "location_id",
                                    "order_date",
                                    "created_by",
                                    "items"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "location_id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-06"
                                    },
                                    "expected_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-13"
                                    },
                                    "created_by": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "qty",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 15000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PO berhasil dibuat"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase/orders/receivable": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Get POs that can be received",
                "operationId": "74bba0d2355023a09d7adceaf09b7a88",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase/orders/{id}": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Get PO details",
                "operationId": "3716e519d4c0d759c699604cfecf3e28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "PO tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Delete a draft PO",
                "operationId": "146d3bcb25d22686d52715a0194a6645",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PO berhasil dihapus"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase/orders/{id}/approve": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Approve a draft PO (set to OPEN)",
                "operationId": "7a02383308b378844d9d24a71035d929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PO berhasil diapprove"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase/orders/{id}/receive": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Receive items from PO (creates Inbound GRN)",
                "operationId": "5806327e160247b0f320fd0e13043418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "received_by",
                                    "items"
                                ],
                                "properties": {
                                    "received_by": {
                                        "type": "string",
                                        "example": "warehouse_staff"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "purchase_order_item_id",
                                                "qty"
                                            ],
                                            "properties": {
                                                "purchase_order_item_id": {
                                                    "type": "string",
                                                    "example": 1
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receive berhasil, Inbound GRN dibuat"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Cancel a PO",
                "operationId": "7e2fe8caebeb24ff55549344cfaba214",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PO berhasil dibatalkan"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}